update according to client requirements, add caculation rules, wait for api fixing and deployment
